--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,6 @@
+# Reference implementations
+reference/
+
 # Compiled files
 *.o
 *.mod