… CLAUDE.md
… CONVOLUTION_MILESTONE.md
… DEVELOPMENT_PATTERNS.md
… MOCK_TO_REAL_CLEANUP.md
… PARSER_MILESTONE.md
… README.md
… REFERENCE_IMPLEMENTATIONS.md
… RESTRUCTURING_SUMMARY.md
… SUMMARY.md
… benchmarks/gpu_baseline_results.md
… benchmarks/gpu_execution_summary.md
… bin/test_conv_gpu_real
… bin/test_convolution_profiles
… bin/test_fortran_c_string
… bin/test_glsl_debug
… bin/test_gpu_conv_egl_combined
… bin/test_gpu_conv_hardcoded
… bin/test_gpu_conv_simple
… bin/test_gpu_egl
… bin/test_gpu_minimal
… bin/test_gpu_minimal2
… docs/ADAPTIVE_PARAMS_STATUS.md
… docs/AMDGPU_SHADER_EXECUTION.md
… docs/BENCHMARKING_PLAN.md
… docs/CURRENT_STATE.md
… docs/DEVLOG_2024_12_ADAPTIVE_PARAMS.md
… docs/DYNAMIC_SHADER_APPROACH.md
… docs/FEATURE_ROADMAP.md
… docs/FORTRAN_PARSER_DESIGN.md
… docs/FORTRAN_SHADER_MVP.md